+++ src/main/java/com/hamza/fruitsappbackend/service/service_impl/UserServiceImpl.java
@@ -6,9 +6,11 @@ import com.hamza.fruitsappbackend.exception.UserNotFoundException;
 import com.hamza.fruitsappbackend.exception.RoleNotFoundException;
 import com.hamza.fruitsappbackend.repository.UserRepository;
 import com.hamza.fruitsappbackend.repository.RoleRepository;
+import com.hamza.fruitsappbackend.security.JwtTokenProvider;
 import com.hamza.fruitsappbackend.service.UserService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -27,24 +29,47 @@ public class UserServiceImpl implements UserService {
     private final RoleRepository roleRepository;
     private final ModelMapper modelMapper;
     private final PasswordEncoder passwordEncoder;
+    private final JwtTokenProvider jwtTokenProvider;
 
     @Autowired
     public UserServiceImpl(UserRepository userRepository, RoleRepository roleRepository,
-                           ModelMapper modelMapper, PasswordEncoder passwordEncoder) {
+                           ModelMapper modelMapper, PasswordEncoder passwordEncoder, JwtTokenProvider jwtTokenProvider) {
         this.userRepository = userRepository;
         this.roleRepository = roleRepository;
         this.modelMapper = modelMapper;
         this.passwordEncoder = passwordEncoder;
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
+
+    private void checkUserOrAdminRole(String token, Long userId) {
+        String username = jwtTokenProvider.getUserNameFromToken(token);
+        User user = userRepository.findByEmail(username)
+                .orElseThrow(() -> new AccessDeniedException("User not found"));
+
+        if (!user.getId().equals(userId) && user.getRoles().stream()
+                .noneMatch(role -> role.getName().equals("ROLE_ADMIN"))) {
+            throw new AccessDeniedException("You do not have the necessary permissions to perform this operation");
+        }
+    }
+
+    private void checkAdminRole(String token) {
+        String username = jwtTokenProvider.getUserNameFromToken(token);
+        User user = userRepository.findByEmail(username)
+                .orElseThrow(() -> new AccessDeniedException("User not found"));
+
+        if (user.getRoles().stream()
+                .noneMatch(role -> role.getName().equals("ROLE_ADMIN"))) {
+            throw new AccessDeniedException("You do not have the necessary permissions to perform this operation");
+        }
     }
 
     @Override
     public UserDTO saveUser(UserDTO userDTO) {
         User user = modelMapper.map(userDTO, User.class);
-
         user.setPassword(passwordEncoder.encode(user.getPassword()));
 
         if (user.getRoles() == null) {
-            user.setRoles(new HashSet<>());
+            throw new RuntimeException("Roles cannot be null");
         }
 
         if (userDTO.getRoles() != null && !userDTO.getRoles().isEmpty()) {
@@ -54,14 +79,16 @@ public class UserServiceImpl implements UserService {
                     .collect(Collectors.toSet());
             user.setRoles(roles);
         }
+
         mapAndSetRelatedEntities(userDTO, user);
         User savedUser = userRepository.save(user);
         return modelMapper.map(savedUser, UserDTO.class);
     }
 
-
     @Override
-    public Optional<UserDTO> getUserById(Long id) {
+    public Optional<UserDTO> getUserById(Long id, String token) {
+        checkUserOrAdminRole(token, id);
+
         return userRepository.findById(id)
                 .map(user -> modelMapper.map(user, UserDTO.class))
                 .or(() -> {
@@ -70,23 +97,35 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Optional<UserDTO> getUserByEmail(String email) {
-        return userRepository.findByEmail(email)
-                .map(user -> modelMapper.map(user, UserDTO.class))
-                .or(() -> {
-                    throw new UserNotFoundException("email", email);
-                });
+    public Optional<UserDTO> getUserByEmail(String email, String token) {
+        String username = jwtTokenProvider.getUserNameFromToken(token);
+        User requestingUser = userRepository.findByEmail(username)
+                .orElseThrow(() -> new AccessDeniedException("User not found"));
+
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new UserNotFoundException("email", email));
+
+        if (!requestingUser.getEmail().equals(email) && requestingUser.getRoles().stream()
+                .noneMatch(role -> role.getName().equals("ROLE_ADMIN"))) {
+            throw new AccessDeniedException("You do not have the necessary permissions to perform this operation");
+        }
+
+        return Optional.of(modelMapper.map(user, UserDTO.class));
     }
 
     @Override
-    public List<UserDTO> getAllUsers() {
+    public List<UserDTO> getAllUsers(String token) {
+        checkAdminRole(token);
+
         return userRepository.findAll().stream()
                 .map(user -> modelMapper.map(user, UserDTO.class))
                 .collect(Collectors.toList());
     }
 
     @Override
-    public UserDTO updateUser(UserDTO userDTO) {
+    public UserDTO updateUser(UserDTO userDTO, String token) {
+        checkUserOrAdminRole(token, userDTO.getId());
+
         User existingUser = userRepository.findById(userDTO.getId())
                 .orElseThrow(() -> new UserNotFoundException("id", userDTO.getId().toString()));
 
@@ -115,7 +154,9 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUserById(Long id) {
+    public void deleteUserById(Long id, String token) {
+        checkUserOrAdminRole(token, id);
+
         if (!userRepository.existsById(id)) {
             throw new UserNotFoundException("id", id.toString());
         }
@@ -123,9 +164,19 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUserByEmail(String email) {
+    public void deleteUserByEmail(String email, String token) {
+        String username = jwtTokenProvider.getUserNameFromToken(token);
+        User requestingUser = userRepository.findByEmail(username)
+                .orElseThrow(() -> new AccessDeniedException("User not found"));
+
         User user = userRepository.findByEmail(email)
                 .orElseThrow(() -> new UserNotFoundException("email", email));
+
+        if (!requestingUser.getEmail().equals(email) && requestingUser.getRoles().stream()
+                .noneMatch(role -> role.getName().equals("ROLE_ADMIN"))) {
+            throw new AccessDeniedException("You do not have the necessary permissions to perform this operation");
+        }
+
         userRepository.delete(user);
     }
